refactor(kernel): introduce generic arch interface
