+++ src/kernel/arch/generic/init.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#if defined(__i386__) || defined(__x86_64__)
+    #include <kernel/arch/x86/init.h>
+#endif
