+++ src/kernel/arch/x86/init.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void
+arch_init();
